[iOS][byot] Add functions to retrieve first run info
This CL adds methods that synchronously returns the time a user
completes their first-run experience, by retrieving the file info of the
first run sentinel file and find the creation time.

Since file info retrieval, as a thread blocking task, has to be done on
a separate thread, a separate method that loads the sentinel info is
added and invoked as soon as the app launches and threads are created.
This way, the creation time retrieval could be done synchronously on the
main thread.

Fixed: 1418111
Change-Id: I01bc5f7aabbd2e24d64a9db58461afcbee68f290
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4307097
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1114079}
NOKEYCHECK=True
GitOrigin-RevId: d45c23691aacbb561a48e32a9a91e8a260491e05