[ios] Extract the URL parameters when the page use blob
When the page uses blob url for the download, the original url
looks like "blob:..." where "..." is the url that includes the
parameters.

Add some code in ARQuickLookTabHelper to try to extract the
parameters from the "blob:..." url. This is a best effort and
may break at any point if/when WKWebView decide to change how
the url is communicated to the application.

Fixed: 1341660
Change-Id: I3abdae9217a8ae9470ab219d06869ff83336dc8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4316384
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Joemer Ramos <joemerramos@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1114509}
NOKEYCHECK=True
GitOrigin-RevId: d858d24ceaa0e2630556996fe81f024ac7dc841e