Shopping service no longer uses the bookmark's price tracking bit
This patch updates the utils and tests in the shopping service to no
longer depend on the price tracking bit in the power bookmark meta.
Instead, the new async api in the subscriptions manager is considered
to be the source of truth. This entailed updating the existing
dependents to work on an async api and updating the relevant tests.

This patch needs to be submitted with the android-side of this work
found here: crrev.com/c/4288121

Bug: b:265329740
Change-Id: I18e7f2cceb0c4ce974fdc92bbf3b8dec09e61f9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4278641
Reviewed-by: Mei Liang <meiliang@chromium.org>
Reviewed-by: Emily Shack <emshack@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Zhiyuan Cai <zhiyuancai@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Yue Zhang <yuezhanggg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1114643}
NOKEYCHECK=True
GitOrigin-RevId: 8d8a75b4c52287eeefc6375ba505fb7b6a6d0aff