[iOS] rename _rootNode to displayedFolderNode.
Bookmark Home had a iv named _rootNode.
This is confusing, because there is one root node by bookmark model.
In particular, the method isDisplayingBookmarkRoot does not refer
_rootNode.
I propose to change the name to displayedFolderNode, to
indicate simply  that we display the content of this folder.

Since there was already a public setter, I changed it to be a property.

Bug: None
Change-Id: Id859aa7698700f116105d77386e61da4d823b49f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4319405
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1114750}
NOKEYCHECK=True
GitOrigin-RevId: 84b3a103dd33d71ffeb08558effc8cf7ab51cafe