[ios] Fix app close refresh can have NTP visible
ForegroundFeedNotVisible refreshes should not refresh when the NTP
is visible. However, app close refreshes should be able to refresh
even when the NTP is visible.

Use a separate variable to track whether the user has sufficiently
engaged with the feed. Reset this variable whenever the feed is
refreshed since the user has not engaged with the new content.

Bug: 1418430
Change-Id: Ic561dc7f08f49b089dced071fe54f404226f3a85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4327456
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: edchin <edchin@google.com>
Cr-Commit-Position: refs/heads/main@{#1115907}
NOKEYCHECK=True
GitOrigin-RevId: a4ca524ba6ec8b9abc8eb0ce8e7d28aa6c4ec047