[iOS] Add Last Password Checkup section in PasswordCheckupViewController
This patch adds the Password Checkup section (composed of the timestamp
cell and the "Check Again" button) in the
PasswordCheckupViewController's table view. It also updates the unit
tests to verify the content and state of this new section.

Mocks:
https://www.figma.com/file/0ZWgfn1GhpwSLfCbqMWRjt/Bling-Password-Checkup?node-id=652%3A107555&t=LBdbJWv3bX7PyBgR-4

UI Deck:
https://docs.google.com/presentation/d/1FUvWpYKB4mVWZD4HyN1dwHV8ujI-NZwQnhs5lZJFnHc/edit?pli=1#slide=id.g19fe51cde0e_0_163

Bug: 1406540
Change-Id: I9a63a66da4effd91dff9b91dede864a4b97761e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4317803
Commit-Queue: Noémie St-Onge <noemies@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Ernesto Izquierdo Clua <eic@google.com>
Cr-Commit-Position: refs/heads/main@{#1115924}
NOKEYCHECK=True
GitOrigin-RevId: 3c8ceccb6573afbba5688850757b30ef9e4d01b5