Ensure sync end notification is sent
In EGTests with signin in, especially the TabGridTestCase, the UI layer
is not notified of the end of the sync, and therefore is not updated.
This CL adds a call to notify_foreign_session_updated_cb_ in
StartLocalSessionEventHandler() to ensure that the notification is
always triggered.

See bug's description for more details: https://crbug.com/1422634.

Fixed: 1422634
Change-Id: Id4da8cbd5b4ddba7d9d49e2ed4f40e8edcfee78f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4321485
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Menghan Yang <myuu@google.com>
Cr-Commit-Position: refs/heads/main@{#1116370}
NOKEYCHECK=True
GitOrigin-RevId: 57a122bc9a5706c49ac1e120181d4101f9ed836c