Suppress debug logging when posting to ResultDB
The python library which sends test results to resultdb via http post also creates a debug log message for each request. This clutters our logs. We had previously suppressed these logs by setting the minimum log level to "warning" on the logger created by the "requests" python module, but that either never worked or since then the logging has moved to "urllib3.connectionpool".

Perhaps a larger issue is that we're setting the minimum log level for ALL loggers to debug in the run.py main method. Maybe we'd want to consider bumping that up to info if not warning.

Tryjob which seems to have become unassociated with this CL:
https://ci.chromium.org/ui/p/chromium/builders/try/ios-simulator-noncq/1926/overview

Bug: 1423933
Change-Id: I6622b779d6ab950ecb90e0f24a7ffa33a4724d8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4332288
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Will Yeager <wyeager@google.com>
Cr-Commit-Position: refs/heads/main@{#1116452}
NOKEYCHECK=True
GitOrigin-RevId: 50951bf75c3429b4bd5e8db0e554458f67a8e76d