Add FET rules for other iOS promos
This is in preparation for migrating the PromosManager to use the FET
as the backend for tracking events. Now, the promos all have rules in
the feature engagement tracker for when the manager starts reading
those rules. The manager will also do promo-time querying of any
checks that cannot be encoded into a rule (e.g. "is the credential
provider extension enabled").

Bug: 1411497
Change-Id: I3657288988cb3d08105ac908f12073209446bfb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4321946
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#1116558}
NOKEYCHECK=True
GitOrigin-RevId: e241d7862470ddec82b015e3d2121b8e4dd9a711