[iOS] Fix title label line height multiple in Incognito interstitial
In the Arabic translation, the diacritics in one of the words is mixed
up with the previous line. This CL fixes this issue by increasing the
line height multiple of the title label in the Incognito interstitial.

It does so by setting the `attributedText` property of the title label
to a customized attributed string. It is synthesized from the title
label once the latter has been set up by PromoStyleViewController, and
customized to change the line height multiple.

Fixed: 1422896
Change-Id: Ie0ae1a61f64445dfb5ef5f81c11361877709fff2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4320605
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1116910}
NOKEYCHECK=True
GitOrigin-RevId: c9ffd65b9b4eecece7e56bf9031e0e8c6b3b98d5