[iOS/B4P] Offer password generation when PASSWORDS data type is active.
Allow password generation for opted in users as well as synced users.
The new condition includes both cases; if sync is enabled, or if opted
in account storage.

Note: to ensure that the generated password will actually get saved in
the account, password generation should only be offered if the runtime
state is active.

Bug: 1392699
Change-Id: I774af1e977949909ced69feadbef01c431f7fe0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4337500
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Jood Hajeer <jood@google.com>
Reviewed-by: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#1116926}
NOKEYCHECK=True
GitOrigin-RevId: f8302f9476bebf1c629a1fc1b86b1631394c1d1c