[ios] Add a NewTabPageComponentFactory to instantiate components
The NTPCoordinator instantiates certain components when it is started.
During unit testing, it is not possible to inject dependencies because
the code that instantiates these components is internal to the
coordinator. This change creates a NewTabPageComponentFactoryProtocol
which the coordinator can use to instantiate these components. The
protocol is concretely implemented by NewTabPageComponentFactory.

This will allow us to create a test version of the factory protocol to
be used in unit tests, which will in turn allow us to inject test
versions of the components.

Change-Id: Ibb988f0209407ffa26fca02ef7de571de63f8892
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4327354
Commit-Queue: Scott Yoder <scottyoder@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1117344}
NOKEYCHECK=True
GitOrigin-RevId: 3baf596af2e9e5624a36000ed080fa41c9a8b38b