Add AutofillLeipzigFeatureCompatibilityLaunch to field trial config
crrev.com/c/4335536 tried to introduce AutofillLeipzigFeatureCompatibilityLaunch in field_testing_config.json but it got reverted due to test failures in iOS [1].

This CL makes adapts those tests to be compatible with the AutofillAccountProfilesUnionView feature. The feature is disabled for testDeletionOfAddressProfile and testAutofillProfileEditing since they are either a duplicate test or incompatible with the feature.

[1]: https://ci.chromium.org/ui/p/chromium/builders/ci/ios-simulator-noncq/32590/overview

Bug: 1348294
Change-Id: Ic8e25be14441c43da8e38f5f6b6d0fc9e38ccc61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4339929
Commit-Queue: Vidhan Jain <vidhanj@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1117493}
NOKEYCHECK=True
GitOrigin-RevId: 14254bf32c39790113837fccb3eb3f1af36b4e27