Add demo option for Inactive Tabs flag
This CL adds a new enum value to the Inactive Tabs threshold flag
parameter. It is not meant to be ever used in a Finch config, but can
be set locally in chrome://flags.
With this demo mode, tabs are considered inactive after one minute, and will be
moved the inactive tabs section on the next launch.

Fixed: 1424772
Change-Id: I41da74b262073c5431fce518ae9441a515c6af28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4341069
Commit-Queue: Louis Romero <lpromero@google.com>
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Auto-Submit: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1117560}
NOKEYCHECK=True
GitOrigin-RevId: 56abd90ee8b76b2d49339be8426f1f25daa41ef6