[Password Checkup] Group password issues.
This patch creates a new data model for the Password Issues surface
as part of the changes needed for the Password Checkup feature.

Password Checkup adds checks for weak and reused passwords, on top
of the existing compromised password check. The Password Issues page
needs to change to accommodate the new types of passwords issues.

The changes are:
 - New data model PasswordIssueGroup: Reused passwords are displayed
together with a text header on top. The new data model is a list of such
 groups. Compromised and reused passwords are simply passed in one group
without header.

 - Adding Grouping logic for Reused passwords.

 - Unit tests for the new logic and fixing breaking changes in other
tests.

UI Deck:
Compromised:
https://docs.google.com/presentation/d/1FUvWpYKB4mVWZD4HyN1dwHV8ujI-NZwQnhs5lZJFnHc/edit#slide=id.g19fe51cde0e_0_230

Weak and reused:
https://docs.google.com/presentation/d/1FUvWpYKB4mVWZD4HyN1dwHV8ujI-NZwQnhs5lZJFnHc/edit#slide=id.g19fe51cde0e_0_335

Bug: 1406540
Change-Id: Ie9eec4f60528e29ffd6b6add6c82f99bc9a81e02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4334689
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Ernesto Izquierdo Clua <eic@google.com>
Cr-Commit-Position: refs/heads/main@{#1117802}
NOKEYCHECK=True
GitOrigin-RevId: 7f93742cec067bdff846365a8c572799b3903b12