[ios] Refactor new_tab_page_uma
new_tab_page_uma::RecordAction does not need the browserState, because
it is only used to check if in incognito mode, so a bool can be passed
with this information, instead of passing the whole browserState
object.

Note that new_tab_page_uma::RecordAction is used also in ios_internal,
so at the moment both solutions coexist, this is only temporary.

Bug:1424397

Change-Id: Id89e281e27b5ec22c3545477fcdca2d5b2b50e06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4335300
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Federica Germinario <fedegermi@google.com>
Cr-Commit-Position: refs/heads/main@{#1118019}
NOKEYCHECK=True
GitOrigin-RevId: 5124ac553ab0a922577230dd50550ae20d38d322