Delete local-state google group information when a profile is deleted
This registers a ProfileManagerObserver to catch when a profile has
been marked for permanent deletion (on multi-profile platforms), and
calls through to the GoogleUpdateService to delete the key
corresponding to that profile.

Sending for early review on the code structure (in particular adding
the observer in ChromeVariationsServiceClient), ahead of unit tests
and live testing.

Bug: b/264838828
Change-Id: I1df9dd94e8c117dfd107c8eb85a357f25076993a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4283097
Reviewed-by: Peter Beverloo <peter@chromium.org>
Auto-Submit: James Lee <ljjlee@google.com>
Reviewed-by: Rayan Kanso <rayankans@chromium.org>
Commit-Queue: Peter Beverloo <peter@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1118026}
NOKEYCHECK=True
GitOrigin-RevId: 00b973d9bfbf742f506fcbe2fbbc650a822cc8ec