[ios] Update ios/web_view/tools/build.py to use host_cpu for simulator
When building the framework for simulator, use the same architecture
as the host. This is cleanup in preparation for the removal of the
support for building universal binary from Chromium build rules.

If any one needs to build an universal binary for this framework, it
will have to be updated to build all slices in different build dir
and then merge the binary using lipo manually.

Bug: 1337780
Change-Id: I44d390c12dc7682d9dccaf609604d71ebd7fd34b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4345269
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: John Wu <jzw@chromium.org>
Commit-Queue: John Wu <jzw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1118346}
NOKEYCHECK=True
GitOrigin-RevId: 068492bf5bd4c4963b588bdb47cc80fc40a788a7