[iOS] Automatically start a password check when Password Manager opens
When opening the Password Manager, a password check would sometimes be
automatically triggered depending on the entry point. For example, a
check would be triggered when opening the Password Manager from the
Overflow Menu, but not when opened from the Settings. This patch makes
the password check automatically trigger, no matter the entry point when
the Password Checkup feature is enabled.

Bug: 1406540
Change-Id: I4ab1e667126f308e683e1dfb76c509832cfa65d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4311332
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Noémie St-Onge <noemies@google.com>
Cr-Commit-Position: refs/heads/main@{#1118651}
NOKEYCHECK=True
GitOrigin-RevId: da4b28375087f61c6ff7d25aca31b41cb1378e7d