[ios] Error dot for identity errors on Settings dest in overflow menu
Put the destinations with a badge in a specific section of the
carousel to catch the user's attention with the following logic:

1) Destinations with error badges are put before the ones with
non-error badges.

2) Destinations with any badge are put before the new destinations.

3) Destinations that are new and have a badge are put before the
destinations that have a badge of the same level of importance.
Destinations with important error badges always come before regardless
of the new destination status.

Fix a bug in meantime where the [] operator would be used to
look for the presence of a destination in the destinations map
instead of using contains(). The [] operator inserts a default value
for the key when there is no match, which could trigger bugs. Hence, []
shouldn't be used to verify the presence of a key.

Mock: http://screen/4iopvfTPTSaADSb
Implementation Screenshot:  http://screen/6xGJdK4qmWAPK9R
Implementation Screencast: http://shortn/_CiLYiVAdEV

Bug: 1420758, 1423345
Change-Id: Ie202f74cdb0a90950dafd49f195c897257c779cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4300421
Reviewed-by: Benjamin Williams <bwwilliams@google.com>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119459}
NOKEYCHECK=True
GitOrigin-RevId: a55d8029d23c80e136167c65723c22d705b36b78