Add ios::AccountBookmarkModelFactory
Adds a factory for BookmarkModel with account storage. The factory
returns null unless kEnableBookmarksAccountStorage feature flag is
enabled. The new factory is hooked up in SyncApiComponentFactoryImpl,
and several TODO's are added in other factories that will need to be
aware of AccountBookmarkModelFactory in the future.

Bug: 1404250
Change-Id: Ida98bcf70f1459675ce17e2345d0914bbc9a4a07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4346652
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119460}
NOKEYCHECK=True
GitOrigin-RevId: 23941830e89e4513adac9bb07c489a37aebf6c70