Move autofill to WebFramesManager::Observer
To prepare for moving autofill to an isolated world, observers need to
always be associated with the correct WebFramesManager, but available
frames reported through the WebStateObserver API are only associated
with the page content world.

This will currently have no effect on the existing call sites as the
same WebFramesManager and frames will be used.

Bug: 1175793

Change-Id: I922c53644b3c0b2669664e5fb4c49ed7e44cc4c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4336705
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119482}
NOKEYCHECK=True
GitOrigin-RevId: 5462e6d9430fe3a562258f6eec5bd019e32e5438