[iOS] Add Insecure Types section in PasswordCheckupViewController
This CL adds the Insecure Types section (composed of the compromised,
reused and weak passwords items) in the PasswordCheckupViewController's
table view. It also updates the unit tests to verify the content of the
items in this new section.

This CL also makes sure that the consumer
(PasswordCheckupViewController) gets notified when it's initialized.

Mocks:
https://www.figma.com/file/0ZWgfn1GhpwSLfCbqMWRjt/Bling-Password-Checkup?node-id=652%3A107555&t=LBdbJWv3bX7PyBgR-4

UI Deck:
https://docs.google.com/presentation/d/1FUvWpYKB4mVWZD4HyN1dwHV8ujI-NZwQnhs5lZJFnHc/edit?pli=1#slide=id.g19fe51cde0e_0_163

Bug: 1406540
Change-Id: Iaa45ed8f184b1c406c3c3ddb8b149c86078f2ea4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4334962
Commit-Queue: Noémie St-Onge <noemies@google.com>
Reviewed-by: Ernesto Izquierdo Clua <eic@google.com>
Cr-Commit-Position: refs/heads/main@{#1119649}
NOKEYCHECK=True
GitOrigin-RevId: 9edc9b64636bb86bf0f9480e6e8f879f1bf4355b