[iOS] Add Browser and baseViewController to SigninPromoViewMediator
To implement the sign-in promo in Bookmarks, SigninPromoViewMediator
needs `browser` and a base view controller to present the sign-in
views.
This patch bring those values `SigninPromoViewMediator`.
The implementation of the sign-in promo is done in a following patch.

Related patches:
=>crrev.com/c/4353591 pipe browser and base view controller
* crrev.com/c/4353591 pipe base view controller in bookmarks
* crrev.com/c/4352830 implement sign-in promo

Bug: 1420997
Change-Id: Idcbfce26deb36db1ca864bead12f398c2829f8e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4352537
Reviewed-by: Petro Akzhygitov <pakzhygitov@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119914}
NOKEYCHECK=True
GitOrigin-RevId: cf7825cbd60710a83ac9085227ee47f9b637b0ed