[ios] FeedStateTracker
This new interface adds a method to know if the feed is visible
in an app close refresh. It also starts to consolidate other methods
around the state of the feed.

The reason we need to know feed visibility is that we found a bug
where the UI update for a new feed model is slow when the user
foregrounds the app directly into the NTP feed.
So we're going to test not doing app close refresh if the feed
is visible while going into the background.

Change-Id: I7e3f23abd212fc4b6a4e4ab580692dd0b59969b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4355258
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: edchin <edchin@google.com>
Cr-Commit-Position: refs/heads/main@{#1120330}
NOKEYCHECK=True
GitOrigin-RevId: e75d85013edbca01b1b4116f84d108dbd78448f5