[ios] Create wrappers for some content concepts
ContentNavigationManager wraps a content::NavigationController
ContentNavigationItem wraps a content::NavigationEntry, and a
ContentNavigationContext wraps a content::NavigationHandle.

The navigation manager is owned by the WebState and it delegates
functionality to its corresponding NavigationController. The
NavigationItem and NavigationContext instances that are vended
are associated with (and owned by) their content:: counterparts
via SharedUserData.

WebState itself is a WebContentsObserver and forwards along to
WebStateObservers.

credit: this CL is predated by a similar sketch from justincohen@.

Bug: 1419001
Change-Id: I96f3c425d4f9a46bc58cc765fcdc27ce4db270ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4341429
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Ian Vollick <vollick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120594}
NOKEYCHECK=True
GitOrigin-RevId: cc112486ca6263679e0fbd8e223db87e3615c1dd