Manipulating Push Notification Feature-level Permissions via Toggle
In a previous CL, the price tracking settings toggle was hooked up to
the prefService, and the toggled display the correct permission status
for the user.

In this CL, the toggle's functionality was extended to be able to change
the permission status for a given push notification enabled feature for
the user and store the new value in the PrefService. In addition, a new
function PushNotificationService::SetPreferences was introduced. This
new function will register the permission changes for the given user to
the push notification server.

In a future CL, the new PushNotificationService::SetPreferences function
will be implemented downstream which will mark the completion of the
issue.

This is a multi-part CL:
  => https://crrev.com/c/4133444
  -  https://crrev.com/i/5272110
  -  https://crrev.com/c/4133646

Bug: 1369616
Change-Id: Iba692f8068a6aff124eee825ec06f19d03623618
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4133444
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Daniel White <danieltwhite@google.com>
Cr-Commit-Position: refs/heads/main@{#1121010}
NOKEYCHECK=True
GitOrigin-RevId: 80f7fdcbc1a0b134dfd1666805588b3b5fd14f32