[iOS] Convert PromosManager to a keyed service
The PromosManager will soon depend on the Feature Engagement Tracker.
That class is a Keyed Service, so the PromosManager also needs to be
one to be able to use this. This doesn't change the behavior today
because iOS only has one regular BrowserState at a time, and there will
be no incognito-specific PromosManager or Feature Engagement Tracker.

Bug: 1408214
Change-Id: Iaa8a415d41db4b529416f7ab7f5e266b53b05625
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4361984
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#1121743}
NOKEYCHECK=True
GitOrigin-RevId: 4e8e5b69f7f3bf357f1f61665c2b79fb0cbbb3e7