[ios] Fix refresh logic re: feed visibility
App close refresh should not happen if the feed is visible while
the app is backgrounding.

The current logic is flawed because the value of `isFeedVisible` is
not correct. It is intended to keep the state of the feed visibility
toggle in the feed header menu. But it is not correctly initialized
inside FeedMetricsRecorder.

Because of the incorrect `isFeedVisible` value, app close refresh
is always incorrectly refreshing.

It is good enough to simply check isNTPVisible.

This CL is part of a multi-step CL. After this CL:
https://crrev.com/i/5654370
https://crrev.com/c/4370625

bug: 1427640
Change-Id: I8b1f2c7931034882909b58ccdaff5e6a9f1eb7f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4368210
Commit-Queue: edchin <edchin@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1121980}
NOKEYCHECK=True
GitOrigin-RevId: 645d60c90b0edb21a9405c2f7f0710ffbf47dbdc