[ios] Silence Overlay Banner Crash and use -started check
This change adds a self.started check before continuing to execute
-finishPresentation logic. This is a guess since it is unclear why
either the banner_view or presentin_view_controller.view is null.

Since the underlying reason has not been found yet, an early return
check is added to silence the issue, but keeps the DCHECK so that
we continue to get pre-stable signals that it is happening.

Fixed: 1215961
Change-Id: Icf1a21f926c89d914502fb7a3b18d8f9c1b16db5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4368842
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1122525}
NOKEYCHECK=True
GitOrigin-RevId: 0be4f41ac234cf39a0299a9e84cadc72b002f928