Add CWVUserContentController message handler API
APIs to add and remove message handlers are currently on CWVWebView.
This is confusing because the handler will be called when the message
is sent by any web view, not only the instance which was used to
register the message.

This CL moves the APIs to CWVUserContentController which more closely
matches WebKit and better represents the scope of which CWVWebViews
can send a message to the registered handler.

Change-Id: If49cca03ebe9989ea1b2fa97024113a8fce2d08a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4362759
Reviewed-by: John Wu <jzw@chromium.org>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Quick-Run: Dan Riley <danielriley@google.com>
Cr-Commit-Position: refs/heads/main@{#1122584}
NOKEYCHECK=True
GitOrigin-RevId: 235fef6044b15884d64753f23629a4cc58babdf4