[ios][feed] Enable background capability at startup for AppCloseBackgroundRefresh
This CL generalizes the code to enable background refresh capability at startup. Currently, it was only used for the background refresh feature.

The new method IsFeedBackgroundRefreshCapabilityEnabled() gates the
capability itself. The capability is enabled if either BackgroundRefresh
or AppCloseBackgroundRefresh is enabled.

Recall that it is done this way because registering for background refreshes must happen early in app initialization and FeatureList is not yet available. So the capability must be enabled for any background
refresh feature.

Bug: 1428446
Change-Id: Iced7761a2ba07cb4913d5c4d877afdba2d8afbee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4375907
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: edchin <edchin@google.com>
Cr-Commit-Position: refs/heads/main@{#1123246}
NOKEYCHECK=True
GitOrigin-RevId: bc110930b3f92ffab7ee40cf734b73d66782fa14