Update gtest_utils to handle spawning test launcher.
When tests fail the stdout is parsed normally by the gtest_utils.
However if the subprocess didn't encounter any fails the stdout
is not written out by the parent process and we need to parse the
[1/1000] TEST (X ms) results to add them as success.

Do not retry for tests run via the spawning test runner.

Bug: 1412835
Change-Id: Ic41cb0a791dc9fef7f55508f0fb7b38e568dadb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4362580
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1123613}
NOKEYCHECK=True
GitOrigin-RevId: 8fb561470e15c0e9c78c2c6534aad2e5e4e2c1f5