[TTFCC] Announce filling payment info
A new method DidFillOrPreviewForm is introduced to the AutofillClient
so it can be notified that a form was filled in.
When payment info is filled into a credit card form on Android, TalkBack
announces "The form was filled in".

Bug: 1427474, 1247698
Change-Id: Ia2f92f09376620ecad381dbb880cbfc36601a326
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4372287
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1123734}
NOKEYCHECK=True
GitOrigin-RevId: 7518f6c0a393a88a2da5d35748c1c13efd2509a2