[iOS] Fix infobar modal reappears on cancellation
This CL adds the PasswordInfobarBannerOverlayRequestCallbackInstaller so
that the callbacks for the password infobar banners are only called
when the request has the same password action as the one passed to the
installer. This new installer prevent the callbacks for the password
infobar banners to be called twice by returning early when the request's
password action doesn't match.

Fixed: 1423339
Change-Id: Id92f87c31b7534ae27b54c3b874110d6409336ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4370423
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Noémie St-Onge <noemies@google.com>
Cr-Commit-Position: refs/heads/main@{#1124234}
NOKEYCHECK=True
GitOrigin-RevId: ea99e727184921519d1ce23ba6825ac41b67f5db