[ios][deps] Add workaround for `ios_..._bundle` templates
The template `ios_app_bundle` and `ios_framework_bundle` only
forward their dependencies to an internal target whose name is
derived from the real target.

This means that if the missing dependency is in a source file
that is listed directly in the `sources` property of one of
those templates, ios/build/tools/update_deps.py was unable to
find the target in the BUILD.gn files and could not fix them.

Add a workaround to compute the real target name from the name
of the internal target generated by the template.

Bug: none
Change-Id: I26d345713184a50b380cb2557940aa7ab0ef7699
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4381974
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Federica Germinario <fedegermi@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1124621}
NOKEYCHECK=True
GitOrigin-RevId: f3025986d7dbe7b7fe7fab8225b5101e8fd27ee9