Implement open password details from anywhere
Part 1 for Issue 142235: Implement long press for the Password Bottom
Sheet.

This CL is just the logic to open the password details view. It is not
called by any line of code in this CL. We are following the same logic
for how we open the Password Manager for manual fallback.

Bug: 1422351
Change-Id: I6728fa3b4c4b23cfcc65a58d37e50a61e368b368
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4367240
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Alexis Hétu <sugoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1124731}
NOKEYCHECK=True
GitOrigin-RevId: 537a2c9f4cdc4bccb40a0af4b018e7ba125e1418