[iOS] Fix bug to move to next field on next in Autofill settings
Before this CL, clicking on "Next" in the keyboard when focused on any field in the profile edit settings did nothing. The logic to move to next text field is implemented in AutofillEditTableViewController. However, when the delegate for the text fields was assigned as AutofillProfileEditTableViewController. This is changed to AutofillSettingsProfileEditTableViewController which extends AutofillEditTableViewController, thus now, on clicking "Next" AutofillEditTableViewController::textFieldShouldReturn is called.

Bug: 1407666,1429593
Change-Id: I9324474a6c4e37bdf732edecdec7d10c1faea62b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4388626
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Vidhan Jain <vidhanj@google.com>
Cr-Commit-Position: refs/heads/main@{#1125302}
NOKEYCHECK=True
GitOrigin-RevId: 73023d158a933fd2968da8386140814ecd5aac19