[iOS] Lower minimum visible percent in a Password Controller EG test
PasswordViewControllerTestCase/testPasswordControllerWhileSearching was
disabled because it was failing on bots. The issue was that the
ManualFallbackPasswordTableViewMatcher was expected to have a minimum
visible percent of 0.6 at the end of the test, but actually got 0.59 on
the bots. This CL re-enables the test and lowers the minimum visible
percent of this matcher to 0.5 to make sure the test passes on all iOS
simulators/devices.

Fixed: 1427909
Change-Id: I8940d98c35baafa646e8f52fa9f5d3570f008ef1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4389234
Commit-Queue: Noémie St-Onge <noemies@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1125344}
NOKEYCHECK=True
GitOrigin-RevId: c7a1d8358dd5cfe7df50a6295a32e8903e013cd2